@@ -5,55 +5,50 @@
   "description": "React.js Hooks to connect your components with Backbone.js models, events and collections.",
   "cdn": "dist/index.umd.js",
   "main": "dist/index.js",
-  "types": "types/index.d.ts",
+  "types": "dist/types/index.d.ts",
   "unpkg": "dist/index.umd.js",
   "module": "dist/index.esm.js",
   "jsdelivr": "dist/index.umd.js",
   "umd:main": "dist/index.umd.js",
   "exports": {
-    ".": [
-      {
-        "import": "./dist/index.mjs",
-        "require": "./dist/index.cjs",
-        "default": "./dist/index.js"
-      },
-      "./dist/index.js"
-    ]
+    ".": {
+      "types": "./dist/types/index.d.ts",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
+    }
   },
   "files": [
     "dist/",
     "types/"
   ],
   "devDependencies": {
-    "@bitty/deferred": "^0.1.0",
+    "@bitty/deferred": "^0.2.0",
+    "@eslint/js": "^9.30.1",
     "@rollup/plugin-commonjs": "^28.0.6",
     "@rollup/plugin-node-resolve": "^16.0.1",
     "@rollup/plugin-terser": "^0.4.4",
-    "@testing-library/react": "^14.0.0",
-    "@types/backbone": "^1.4.15",
-    "@types/eslint": "^8.40.0",
-    "@types/jest": "^29.5.1",
-    "@types/prettier": "^2.7.2",
-    "@types/react": "^18.2.7",
-    "@types/react-dom": "^18.2.4",
-    "@typescript-eslint/eslint-plugin": "^5.59.7",
-    "@typescript-eslint/parser": "^5.59.7",
-    "backbone": "^1.4.1",
-    "eslint": "^8.41.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "prettier": "^2.8.8",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "@rollup/plugin-typescript": "^12.1.4",
+    "@testing-library/react": "^16.3.0",
+    "@types/backbone": "^1.4.23",
+    "@types/eslint": "^9.6.1",
+    "@types/jest": "^30.0.0",
+    "@types/prettier": "^2.7.3",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "backbone": "^1.6.1",
+    "eslint": "^9.30.1",
+    "jest": "^30.0.4",
+    "jest-environment-jsdom": "^30.0.4",
+    "prettier": "^3.6.2",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "rollup": "^4.44.2",
-    "rollup-plugin-typescript2": "^0.36.0",
-    "ts-jest": "^29.1.0",
+    "ts-jest": "^29.4.0",
     "ts-jest-resolver": "^2.0.1",
-    "typedoc": "^0.24.7",
-    "typedoc-plugin-markdown": "^3.15.3",
-    "typescript": "^5.0.4"
+    "typedoc": "^0.28.7",
+    "typedoc-plugin-markdown": "^4.7.0",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.35.1"
   },
   "peerDependencies": {
     "backbone": ">= 1.4",
@@ -62,8 +57,10 @@
   "scripts": {
     "doc": "typedoc src/index.ts",
     "test": "jest",
-    "lint": "eslint \"*/**/*.{ts,js,json}\"",
-    "lint:fix": "eslint \"*/**/*.{ts,js,json}\" --fix",
+    "format": "prettier \"**/*.{json,mjs,ts}\" --write",
+    "format:check": "prettier \"**/*.{json,mjs,ts}\" --check",
+    "lint": "eslint \"**/*.{mjs,ts}\"",
+    "lint:fix": "eslint \"**/*.{mjs,ts}\" --fix",
     "build": "rollup --config ./rollup.config.mjs",
     "prepublishOnly": "npm run doc && npm run lint && npm run test && npm run build"
   },
@@ -81,8 +78,8 @@
   },
   "homepage": "https://github.com/VitorLuizC/react-hooks-for-backbone#readme",
   "dependencies": {
-    "@types/jquery": "^3.5.16",
-    "@types/lodash.isequal": "^4.5.6",
+    "@types/jquery": "^3.5.32",
+    "@types/lodash.isequal": "^4.5.8",
     "lodash.isequal": "^4.5.0"
   }
 }
